3 add testcoverage for customer usecases (#5)
* Upgrade Dependencies and Fixed multiple bugs

* Create Basic tests

* Delete dummy test

+++ src/AKS.Application/UseCases/Customers/Create/CreateCustomer.cs
@@ -8,5 +8,5 @@ public record CreateCustomer(
     string? State,
     string? Country,
     string? ZipCode,
-    long Latitude,
-    long Longitude);
+    double Latitude,
+    double Longitude);
